+++ src/main/java/com/synergyx/trading/controller/StockController.java
@@ -3,8 +3,8 @@ package com.synergyx.trading.controller;
 import com.synergyx.trading.apiPayload.ApiResponse;
 import com.synergyx.trading.dto.stockDetail.StockDetailResponseDTO;
 import com.synergyx.trading.dto.stockSearch.StockSearchResponseDTO;
-import com.synergyx.trading.service.stockDetailService.StockDetailQueryService;
-import com.synergyx.trading.service.stockSearchService.StockSearchQueryService;
+import com.synergyx.trading.service.stockService.detail.StockDetailQueryService;
+import com.synergyx.trading.service.stockService.search.StockSearchQueryService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.tags.Tag;
